Rename and fix PayPal and Klarna
