+++ lib/Checkout/Payments/Request/Source/Apm/RequestPayPalSource.php
@@ -1,16 +1,20 @@
 <?php
 
-namespace Checkout\Payments\Request\Source\Contexts;
+namespace Checkout\Payments\Request\Source\Apm;
 
 use Checkout\Common\PaymentSourceType;
+use Checkout\Payments\BillingPlan;
 use Checkout\Payments\Request\Source\AbstractRequestSource;
 
-class PaymentContextsPaypalSource extends AbstractRequestSource
+class RequestPayPalSource extends AbstractRequestSource
 {
-
     public function __construct()
     {
         parent::__construct(PaymentSourceType::$paypal);
     }
 
+    /**
+     * @var BillingPlan
+     */
+    public $plan;
 }
