+++ lib/Checkout/Payments/Request/Source/Contexts/PaymentContextsPayPalSource.php
@@ -1,20 +1,16 @@
 <?php
 
-namespace Checkout\Payments\Request\Source\Apm;
+namespace Checkout\Payments\Request\Source\Contexts;
 
 use Checkout\Common\PaymentSourceType;
-use Checkout\Payments\BillingPlan;
 use Checkout\Payments\Request\Source\AbstractRequestSource;
 
-class RequestPaypalSource extends AbstractRequestSource
+class PaymentContextsPayPalSource extends AbstractRequestSource
 {
+
     public function __construct()
     {
         parent::__construct(PaymentSourceType::$paypal);
     }
 
-    /**
-     * @var BillingPlan
-     */
-    public $plan;
 }
